todo/requirements.txt
Django==5.2.6
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
psycopg2-binary==2.9.10
requests==2.32.5
gunicorn>=21.2,<22.0